requirements.txt
Pillow
fpdf
pyautogui
numpy
opencv
ocrmypdf
